requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
boto3==1.9.14
python-dotenv==0.10.3